[python] Restructure python/ to src layout for PyPI packaging (#180)
Restructures the `python/` directory to a standard `src` layout and adds
PyPI
packaging scaffolding (reserving the `decimo` package name).

## Changes

### PyPI packaging (`python/pyproject.toml`, `python/README.md`)
- Add `pyproject.toml` with `hatchling` build backend and `src` layout.
- Add `python/README.md` as the PyPI landing page.
- Add `python/src/decimo/py.typed` (PEP 561 marker for type checkers).
- Placeholder wheel `0.1.0.dev0` published to PyPI — name `decimo` is
now reserved.
- `pixi run wheel` task builds the wheel; `python-build` and `twine`
added to pixi dependencies.

### `src` layout restructure
- `python/decimo.py` → `python/src/decimo/__init__.py`
- `python/_decimo.pyi` → `python/src/decimo/_decimo.pyi`
- `python/tests/test_decimo.py` — updated `sys.path` for new layout.
- `pixi run buildpy` — output path updated to
`python/src/decimo/_decimo.so`.
- `.vscode/settings.json` — `extraPaths` updated to `["python/src"]`.
