+++ .gitignore
@@ -29,3 +29,4 @@ local
 *.so
 __pycache__/
 *.pyc
+python/dist/
