+++ python/src/decimo/_decimo.pyi
